Updated player collisions
Had an issue where the player would collide with seemingly invisible walls when colliding with items. Additionally, when approaching an object from below, the player character would collide from their head, which is completely unintended. I've fixed this by creating a smaller collision hitbox within the greater. This will be used purely for collisions with walls and the such, as hits will call for greater precision.