Updated Inventory with tabs
I updated the inventory with seperate inventory tabs. These effectively work as inventories of their own, and are pretty useful for organising different items. At the moment, the main issue is that for some reason tooltips don't seem to display for duplicate items. This could be fixed by implementing item stacking. I'll work on it.