Removal of Controller.py and creation of base entity class with some basic animation
So, I originally had a controller and an animation class to handle movement and animation. I realised this  isn't necessary, and have instead rolled my code for controller and animation into my AnimationEntity class. Reads from a spritesheet and creates image lists, where it selects the correct one depending on the current action. Further information is in the comments for the entity.py files. Next step is to rework player.py and game_world.py  to reflect these changes.